Split util into modules by function
